backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.21.1
sqlalchemy==1.4.49
alembic==1.12.1
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
httpx==0.25.2 